+++ src/main/java/com/studentmanagement/service/StudentServiceImpl.java
@@ -49,10 +49,9 @@ public class StudentServiceImpl implements StudentService {
             Student student = existingStudent.get();
             student.setName(updatedStudent.getName());
             student.setBirthday(updatedStudent.getBirthday());
-            student.setUsername(updatedStudent.getUsername());
-            student.setPassword(updatedStudent.getPassword());
             student.setCredits(updatedStudent.getCredits());
             student.setIdClass(updatedStudent.getIdClass());
+            student.setIndustry(updatedStudent.getIndustry());
             return studentRepository.save(student);
         } else {
             throw new RuntimeException("Student with id " + id + " not found.");
@@ -61,7 +60,11 @@ public class StudentServiceImpl implements StudentService {
 
     @Override
     public void deleteStudent(Integer id) {
-        studentRepository.deleteById(id);
+        if (studentRepository.existsById(id)) {
+            studentRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Student with id " + id + " not found.");
+        }
     }
 
     @Override
